+++ internal/storage/pg/storage_test.go
@@ -1,11 +1,11 @@
 package storage
 
 import (
-	"context"
 	"errors"
 	"testing"
 	"time"
 
+	"github.com/iudanet/yp-metrics-go/internal/retry"
 	"github.com/jackc/pgerrcode"
 	"github.com/jackc/pgx/v5/pgconn"
 	"github.com/stretchr/testify/assert"
@@ -45,42 +45,12 @@ func TestWithRetry(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			err := withRetry(tt.op)
+			err := retry.WithRetry(tt.op)
 			if tt.wantError {
 				assert.Error(t, err)
 			} else {
 				assert.NoError(t, err)
 			}
-		})
-	}
-}
-
-func TestIsRetriableError(t *testing.T) {
-	tests := []struct {
-		name     string
-		err      error
-		expected bool
-	}{
-		{
-			name:     "pg connection error",
-			err:      &pgconn.PgError{Code: pgerrcode.ConnectionException},
-			expected: true,
-		},
-		{
-			name:     "context canceled",
-			err:      context.Canceled,
-			expected: true,
-		},
-		{
-			name:     "non-retriable error",
-			err:      errors.New("some error"),
-			expected: false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			assert.Equal(t, tt.expected, isRetriableError(tt.err))
 		})
 	}
 }
